# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "internet-date API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/quizzes/{quizId}/submit": {
      "post": {
        "tags": [
          "quizzes"
        ],
        "summary": "Submit quiz",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizSubmitBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quizzes/{quizId}": {
      "patch": {
        "tags": [
          "quizzes"
        ],
        "summary": "Update quiz",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizUpdateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizUpdateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quizzes/{quizId}/questions/{questionId}": {
      "patch": {
        "tags": [
          "quizzes"
        ],
        "summary": "Update quiz question",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizQuestionPatchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizQuestionPatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quizzes/{quizId}/questions/{questionId}/options/{optionId}": {
      "patch": {
        "tags": [
          "quizzes"
        ],
        "summary": "Update quiz option",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "optionId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizOptionPatchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizOptionPatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/block": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Block user",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/report": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Report user",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Id": {
        "type": "string",
        "pattern": "^\\d+$"
      },
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "ReportReason": {
        "type": "string",
        "enum": [
          "SPAM",
          "HARASSMENT",
          "IMPERSONATION",
          "NUDITY",
          "HATE",
          "OTHER"
        ]
      },
...
      "QuizSubmitBody": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "object",
            "additionalProperties": true
          },
          "scoreVec": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          }
        },
        "required": [
          "answers"
        ]
      },
      "QuizUpdateBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "QuizUpdateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "updatedAt"
        ]
      },
      "QuizQuestionPatchBody": {
        "type": "object",
        "properties": {
          "prompt": {
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ]
      },
      "QuizQuestionPatchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "prompt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "prompt"
        ]
      },
      "QuizOptionPatchBody": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          }
        },
        "required": [
          "label"
        ]
      },
      "QuizOptionPatchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "label": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "value"
        ]
      },
      "ReportBody": {
        "type": "object",
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/ReportReason"
          },
          "details": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "reason"
        ]
      }
    }
  }
}